+++ CHANGELOG.md
@@ -1,10 +1,61 @@
-## [5.0.18](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/compare/v5.0.17...v5.0.18) (2026-04-14)
+# [6.0.0-oclif4.3](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/compare/v6.0.0-oclif4.2...v6.0.0-oclif4.3) (2026-04-20)
 
 
 ### Bug Fixes
 
+* fix dependencies ([92c58dc](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/commit/92c58dc898af01795945997e2f2ce568968662ce))
 * fix release script ([0609437](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/commit/06094375da448a8ab94a8c340159fc2e202019df))
 * update dependencies and add updates workflow ([543b426](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/commit/543b42671ae266615b9db7ce6db16d68f1d4f465))
+* update dependencies and fix review issues ([f17d30b](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/commit/f17d30b52b9fcc244778e2acf38b7df29bfde433))
+
+
+### Features
+
+* migrate to oclif/core v4 and oclif/test v4 ([77a4114](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/commit/77a411474db7d22775283c4cd971f8adec0fc06e))
+* require node to v22 ([d64a8c1](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/commit/d64a8c1b3c065f0582a19cdc4ef29ee6595a4df8))
+
+
+### BREAKING CHANGES
+
+* require node to v22
+
+# [6.0.0-oclif4.2](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/compare/v6.0.0-oclif4.1...v6.0.0-oclif4.2) (2026-03-27)
+
+
+### Bug Fixes
+
+* update dependencies and fix review issues ([aba947f](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/commit/aba947f4ca743bfccd3774fc6419c8ec01817146))
+
+# [6.0.0-oclif4.1](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/compare/v5.1.0-oclif4.2...v6.0.0-oclif4.1) (2026-03-26)
+
+
+### Features
+
+* require node to v22 ([25951e9](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/commit/25951e99cda1daa023ff30ca1e4fdd38a13db947))
+
+
+### BREAKING CHANGES
+
+* require node to v22
+
+# [5.1.0-oclif4.2](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/compare/v5.1.0-oclif4.1...v5.1.0-oclif4.2) (2026-03-26)
+
+
+### Bug Fixes
+
+* update dependencies ([1df2f03](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/commit/1df2f03448181800269da0726604843dcbaaf76f))
+
+# [5.1.0-oclif4.1](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/compare/v5.0.15...v5.1.0-oclif4.1) (2026-03-24)
+
+
+### Bug Fixes
+
+* add biome ([5eceb91](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/commit/5eceb9161c878f989744a4ed70ba8271d3668097))
+
+
+### Features
+
+* migrate to oclif/core v4 and oclif/test v4 ([50b51da](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/commit/50b51da6216c4d446229c3f0bcbbcc34a48a6d92))
 
 ## [5.0.17](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/compare/v5.0.16...v5.0.17) (2026-03-24)
 
@@ -13,12 +64,17 @@
 
 * update dependencies ([1df2f03](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/commit/1df2f03448181800269da0726604843dcbaaf76f))
 
-## [5.0.16](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/compare/v5.0.15...v5.0.16) (2026-03-24)
+# [5.1.0-oclif4.1](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/compare/v5.0.15...v5.1.0-oclif4.1) (2026-03-24)
 
 
 ### Bug Fixes
 
 * add biome ([5eceb91](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/commit/5eceb9161c878f989744a4ed70ba8271d3668097))
+
+
+### Features
+
+* migrate to oclif/core v4 and oclif/test v4 ([50b51da](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/commit/50b51da6216c4d446229c3f0bcbbcc34a48a6d92))
 
 ## [5.0.15](https://github.com/commercelayer/commercelayer-cli-plugin-checkout/compare/v5.0.14...v5.0.15) (2026-03-13)